alphagalleon-backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
httpx
beautifulsoup4
pandas
numpy
